Core: added TFData
